requirements.txt
Flask
gunicorn
psycopg2-binary
sqlalchemy
